Fixes dyanmodb_cleanup conditional logic
Previously this module was using a private fork of the terraform-aws-lambda module
this change removes that dependancy. In order to do so however, the additional
policy that is passed to the lambda module must include a valid policy.

If the enable flag is set to true this module will be fully integrated with the
step function and provisioned with the proper policy. Otherwise the lamdba function
is still created but without any of its functionality.